+++ RELEASE_NOTES.md
@@ -5,3 +5,9 @@
 ### Version 1.1.0
 - Module updated to handle ref chains for Reads within the input Sample Set
 - Uses an updated version of SetAPI
+
+
+### Version 1.1.4
+- add github action
+- add KBaseMetagenomes.AnnotatedMetagenomeAssembly as a valid input argument that HISAT2 will align against
+- update module to use Python 3